football_focussed/requirements.txt
alembic==1.12.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dotenv==0.21.1
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
